pyproject.toml
[project]
name = "rank-test"
version = "0.1.0"
description = "Compare loss functions for QA retrieval ranking"
readme = "README.md"
authors = [
    { name = "Michael Fine", email = "mfine15@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "kaggle",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.0.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "pytest>=8.3.5",
]

[project.scripts]
rank-test = "rank_test.train_flexible:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
